+++ R/DfReadDataFile.R
@@ -38,7 +38,14 @@
 #'    treatment and reader IDs (i.e., names). Otherwise, treatment 
 #'    and reader IDs in the original data file will be used.
 #' 
-#' @note The \code{"MRMC"} format is deprecated. For non-JAFROC formats four file extensions (\code{.csv}, \code{.txt}, \code{.lrc} and \code{.imrmc}) are possible, all of which are restricted to ROC data. Only the \code{iMRMC} format is actively supported, i.e, files with extension \code{.imrmc}. Other formats (\code{.csv}, \code{.txt}, \code{.lrc}) are deprecated. Such files can still be read by this function and then saved to a JAFROC format file for further analysis within this package. \bold{For non-JAFROC data file formats, the \code{readerID} and \code{modalityID} fields  must be unique integers}.
+#' @note The \code{"MRMC"} format is deprecated. For non-JAFROC formats four file 
+#'    extensions (\code{.csv}, \code{.txt}, \code{.lrc} and \code{.imrmc}) are possible, 
+#'    all of which are restricted to ROC data. Only the \code{iMRMC} format is actively 
+#'    supported, i.e, files with extension \code{.imrmc}. Other formats (\code{.csv}, 
+#'    \code{.txt}, \code{.lrc}) are deprecated. Such files can still be read by this 
+#'    function and then saved to a JAFROC format file for further analysis within this 
+#'    package. \bold{For non-JAFROC data file formats, the \code{readerID} and 
+#'    \code{modalityID} fields  must be unique integers}.
 #' 
 #' @return A dataset with the structure specified in \code{\link{RJafroc-package}}.
 #'