+++ tests/testthat/test-DfSaveDataFile.R
@@ -5,11 +5,11 @@ test_that(contextStr, {
   fn <- paste0(test_path(), "/goodValues361/DfSaveDataFile/dataset05.xlsx")
   if (!file.exists(fn)) {
     warning(paste0("File not found - generating new ",fn))
-    DfSaveDataFile(dataset05, fileName = fn, format = "JAFROC")
+    DfWriteExcelDataFile(dataset05, fileName = fn)
   }
   
   fn1 <- paste0(test_path(), "/goodValues361/DfSaveDataFile/myTemp.xlsx")
-  DfSaveDataFile(dataset05, fileName = fn1, format = "JAFROC")
+  DfWriteExcelDataFile(dataset05, fileName = fn1)
   for (i in 1:3) { # there are 3 worksheets in Excel file
     dfGood <- readWorkbook(fn, i) # check each sheet individually
     dfCurrent <- readWorkbook(fn1, i)    # do:
@@ -29,11 +29,11 @@ test_that(contextStr, {
   fn <- paste0(test_path(), "/goodValues361/DfSaveDataFile/datasetROI.xlsx")
   if (!file.exists(fn)) {
     warning(paste0("File not found - generating new ",fn))
-    DfSaveDataFile(datasetROI, fileName = fn, format = "JAFROC")
+    DfWriteExcelDataFile(datasetROI, fileName = fn)
   }
   
   fn1 <- paste0(test_path(), "/goodValues361/DfSaveDataFile/DfSaveDataFile-roi1.xlsx")
-  DfSaveDataFile(datasetROI, fileName = fn1, format = "JAFROC")
+  DfWriteExcelDataFile(datasetROI, fileName = fn1)
   for (i in 1:3) { # there are 3 worksheets in Excel file
     dfGood <- readWorkbook(fn, i) # check each sheet individually
     dfCurrent <- readWorkbook(fn1, i)    # do: